Try to clear up the line between Juju and MAAS
Avoid talking about services in the context of MAAS, push that down
to the Juju paragraph. Also, include the key word "constraints" when
discussing Juju. I think this will make it easier for users to
find out how to use tags w/ Juju, since "constraints" is a section
in the Juju docs. A keyword search for "tags" on the juju docs main
page, on the other hand, turns up nothing.