Clarify that XPath expressions are associated with a tag
When I first read "MAAS supports XPath expressions in tags", I thought
that meant that this was in place of a tag name. But by playing
with the API, I discovered that the name and the expression are both
properties of a tag. Update the language to make that more immediately clear.